fix: add default subject if subject is empty string
This modifies generate_filename to provide a default subject when an
empty string is passed to the subject parameter, or when subject only
has non-alphanumeric characters. Previously, the function accepted empty
string as subject which generated file names that only have `CURR_DATE`
with a leading underscore (e.g. "_20240703.pdf").