feat: prefix file extension with a dot when generating filenames
Prepend a dot to the ext argument passed to generate_filename if it
does not already have one, and the argument is not an empty string.

Since the ext parameter accepts any strings, this will prevent
generating erroneous file names where the name and the extension are
not separated with a dot. This will also prevent adding a trailing
dot in file names where the user decided not to provide an extension.